Feat: Make command block have a similar interface to Arguments
#break BREAK breaks
#changelog

Changed interface for command block callbacks. When defined, the Arguments
interface can bused to specify the value of the trailing argument that
will be used as the input to the command block function. Updated help
messaging to indicate the arguments a command accepts as well as the value
of the commands own trailing argument (e.g path string) if the command
was configure with `REQUIRES_VALUE` set to true.